chore: fix Sonar issue "Default export names and file names should match" for Swagger component (#2066)
